Move Fonts into separate directory
