+++ CMakeLists.txt
@@ -33,14 +33,12 @@ set(Impacto_Src
     src/spritesheet.cpp
     src/spriteanimation.cpp
     src/sequencedanimation.cpp
-    src/font.cpp
     src/background2d.cpp
     src/mask2d.cpp
     src/character2d.cpp
     src/inputsystem.cpp
     src/voicetable.cpp
     src/animation.cpp
-    src/externalfont.cpp
 
     src/text/backlogpage.cpp
     src/text/text.cpp
@@ -49,6 +47,9 @@ set(Impacto_Src
     src/text/tipspage.cpp
     src/text/typewritereffect.cpp
     src/text/dialoguepage.cpp
+
+    src/text/fonts/externalfont.cpp
+    src/text/fonts/font.cpp
 
     src/effects/wave.cpp
     src/effects/blur.cpp
@@ -403,7 +404,6 @@ set(Impacto_Header
     src/spritesheet.h
     src/spriteanimation.h
     src/sequencedanimation.h
-    src/font.h
     src/background2d.h
     src/mask2d.h
     src/character2d.h
@@ -420,6 +420,8 @@ set(Impacto_Header
     src/text/typewritereffect.h
     src/text/dialoguepage.h
     src/text/processedtextglyph.h
+
+    src/text/fonts/font.h
 
     src/effects/wave.h
     src/effects/blur.h
