+++ src/profile/fonts.h
@@ -1,13 +1,13 @@
 #pragma once
 
-#include "../font.h"
+#include "../text/fonts/font.h"
 #include "../util.h"
 #include <ankerl/unordered_dense.h>
 
 namespace Impacto {
 namespace Profile {
 
-inline ankerl::unordered_dense::map<std::string, Font*, string_hash,
+inline ankerl::unordered_dense::map<std::string, Fonts::Font*, string_hash,
                                     std::equal_to<>>
     Fonts;
